Reavaliação dos models. Criação de alguns repositories e redefinição de um service.

+++ src/main/java/com/ufcg/si1/service/EspecialidadeService.java
@@ -22,7 +22,7 @@ public interface EspecialidadeService {
     void insere(Especialidade esp)throws Rep,
             ObjetoJaExistenteException;
 
-    boolean existe(int codigo);
+    boolean existe(long codigo);
 
     Especialidade findById(long id);
 }
